refactor(plugins): split manifest normalization
