+++ src/plugins/package-manifest.ts
@@ -1,0 +1,197 @@
+import { normalizeOptionalString } from "../../packages/normalization-core/src/string-coerce.js";
+import type { ChannelSetupMetadata } from "../channels/plugins/setup-contract.js";
+import { MANIFEST_KEY } from "../compat/legacy-names.js";
+import { isRecord } from "../utils.js";
+import type { PluginManifestChannelCommandDefaults } from "./manifest-types.js";
+
+/** package.json OpenClaw metadata used for plugin setup and catalog discovery. */
+type PluginPackageChannelApprovalFlag = "native";
+
+export type PluginPackageChannel = {
+  id?: string;
+  label?: string;
+  selectionLabel?: string;
+  detailLabel?: string;
+  docsPath?: string;
+  docsLabel?: string;
+  blurb?: string;
+  order?: number;
+  aliases?: readonly string[];
+  preferOver?: readonly string[];
+  systemImage?: string;
+  selectionDocsPrefix?: string;
+  selectionDocsOmitLabel?: boolean;
+  selectionExtras?: readonly string[];
+  markdownCapable?: boolean;
+  /** Closed manifest flags for approval behavior available before the channel runtime loads. */
+  approvalFlags?: readonly PluginPackageChannelApprovalFlag[];
+  exposure?: {
+    configured?: boolean;
+    setup?: boolean;
+    docs?: boolean;
+  };
+  quickstartAllowFrom?: boolean;
+  forceAccountBinding?: boolean;
+  preferSessionLookupForAnnounceTarget?: boolean;
+  commands?: PluginManifestChannelCommandDefaults;
+  configuredState?: {
+    specifier?: string;
+    exportName?: string;
+    env?: {
+      allOf?: readonly string[];
+      anyOf?: readonly string[];
+    };
+  };
+  persistedAuthState?: {
+    specifier?: string;
+    exportName?: string;
+  };
+  doctorCapabilities?: PluginPackageChannelDoctorCapabilities;
+  /** Typed, serializable setup fields available before plugin runtime load. */
+  setup?: ChannelSetupMetadata;
+  /** @deprecated Use setup.fields. */
+  cliAddOptions?: readonly PluginPackageChannelCliOption[];
+};
+
+export type PluginPackageChannelDoctorCapabilities = {
+  dmAllowFromMode?: "topOnly" | "topOrNested" | "nestedOnly";
+  groupModel?: "sender" | "route" | "hybrid";
+  groupAllowFromFallbackToAllowFrom?: boolean;
+  warnOnEmptyGroupSenderAllowlist?: boolean;
+};
+
+export type PluginPackageChannelCliOption = {
+  flags: string;
+  negatedFlags?: string;
+  description: string;
+  defaultValue?: boolean | string;
+  valueType?: "int" | "list";
+};
+
+export type PluginPackageInstall = {
+  clawhubSpec?: string;
+  npmSpec?: string;
+  localPath?: string;
+  defaultChoice?: "clawhub" | "npm" | "local";
+  minHostVersion?: string;
+  expectedIntegrity?: string;
+  allowInvalidConfigRecovery?: boolean;
+  requiredPlatformPackages?: string[];
+};
+
+type OpenClawPackageStartup = {
+  /**
+   * Opt-in for channel plugins whose `setupEntry` fully covers the gateway
+   * startup surface needed before the server starts listening.
+   */
+  deferConfiguredChannelFullLoadUntilAfterListen?: boolean;
+};
+
+type OpenClawPackageSetupFeatures = {
+  configPromotion?: boolean;
+  legacyStateMigrations?: boolean;
+  legacySessionSurfaces?: boolean;
+};
+
+type OpenClawPackageCompat = {
+  pluginApi?: string;
+  minGatewayVersion?: string;
+};
+
+export type OpenClawPackageBuild = {
+  bundledDist?: boolean;
+  openclawVersion?: string;
+  pluginSdkVersion?: string;
+};
+
+export type OpenClawPackageManifest = {
+  extensions?: string[];
+  runtimeExtensions?: string[];
+  setupEntry?: string;
+  runtimeSetupEntry?: string;
+  setupFeatures?: OpenClawPackageSetupFeatures;
+  plugin?: {
+    id?: string;
+    label?: string;
+  };
+  channel?: PluginPackageChannel;
+  compat?: OpenClawPackageCompat;
+  install?: PluginPackageInstall;
+  startup?: OpenClawPackageStartup;
+  build?: OpenClawPackageBuild;
+};
+
+export const DEFAULT_PLUGIN_ENTRY_CANDIDATES = [
+  "index.ts",
+  "index.js",
+  "index.mjs",
+  "index.cjs",
+] as const;
+
+export type PackageExtensionResolution =
+  | { status: "ok"; entries: string[] }
+  | { status: "missing"; entries: [] }
+  | { status: "empty"; entries: [] }
+  | { status: "invalid"; entries: []; error: string };
+
+type ManifestKey = typeof MANIFEST_KEY;
+
+export type PackageManifest = {
+  name?: string;
+  version?: string;
+  description?: string;
+  dependencies?: Record<string, string>;
+  optionalDependencies?: Record<string, string>;
+} & Partial<Record<ManifestKey, OpenClawPackageManifest>>;
+
+export function getPackageManifestMetadata(
+  manifest: PackageManifest | undefined,
+): OpenClawPackageManifest | undefined {
+  if (!manifest) {
+    return undefined;
+  }
+  return manifest[MANIFEST_KEY];
+}
+
+export function resolvePackageExtensionEntries(
+  manifest: PackageManifest | undefined,
+): PackageExtensionResolution {
+  const rawOpenClaw = manifest?.[MANIFEST_KEY] as unknown;
+  if (rawOpenClaw === undefined || rawOpenClaw === null) {
+    return { status: "missing", entries: [] };
+  }
+  if (!isRecord(rawOpenClaw)) {
+    return {
+      status: "invalid",
+      entries: [],
+      error: "package.json openclaw must be an object",
+    };
+  }
+  const raw = rawOpenClaw.extensions;
+  if (raw === undefined || raw === null) {
+    return { status: "missing", entries: [] };
+  }
+  if (!Array.isArray(raw)) {
+    return {
+      status: "invalid",
+      entries: [],
+      error: "package.json openclaw.extensions must be an array",
+    };
+  }
+  const entries: string[] = [];
+  for (const [index, entry] of raw.entries()) {
+    const normalized = normalizeOptionalString(entry);
+    if (!normalized) {
+      return {
+        status: "invalid",
+        entries: [],
+        error: `package.json openclaw.extensions[${index}] must be a non-empty string`,
+      };
+    }
+    entries.push(normalized);
+  }
+  if (entries.length === 0) {
+    return { status: "empty", entries: [] };
+  }
+  return { status: "ok", entries };
+}
